# jobs: avoid CTEs in crdb_internal.system_jobs query