@rafiss 2024-05-08
it looks like the roundtrips went down, but the latency went up. perhaps the query plan is actually less computationally efficient, even though it performs less I/O? i'd guess there's some threshold at which latency is actually better the new way, and maybe in a multiregion cluster it could matter even more.

```
goos: darwin
goarch: arm64
                               │ old-bench.txt │            new-bench.txt             │
                               │    sec/op     │    sec/op     vs base                │
Jobs/crdb_internal.system_jobs     167.0m ± 6%   292.7m ± 16%  +75.27% (p=0.000 n=10)

                               │ old-bench.txt │           new-bench.txt            │
                               │  roundtrips   │ roundtrips  vs base                │
Jobs/crdb_internal.system_jobs      8.000 ± 0%   3.000 ± 0%  -62.50% (p=0.000 n=10)

                               │ old-bench.txt │             new-bench.txt             │
                               │     B/op      │     B/op       vs base                │
Jobs/crdb_internal.system_jobs    58.55Mi ± 2%   105.99Mi ± 4%  +81.03% (p=0.000 n=10)

                               │ old-bench.txt │            new-bench.txt             │
                               │   allocs/op   │  allocs/op    vs base                │
Jobs/crdb_internal.system_jobs     905.9k ± 2%   1595.0k ± 3%  +76.06% (p=0.000 n=10)
```