@stevendanna 2024-05-08
Yeah, I think where we are hoping this helps substantially is in the case where we have a job_type or job_status filter that filters out a high percentage of rows.  I'll add a test that captures whether it does or not. It seems to in some local testing.