@stevendanna 2024-05-15
Part of what is going on here is that the current tests are using query plans that are created in the absence of stats. If I add an `ANALYZE system.jobs` and `ANALYZE system.job_info` into the setup statements, many of the existing queries choose plans that also have a smaller number of round-trips.  

I'll add some commits to update the existing tests to better reflect reality so we can compare against that.

My believe here is that we should really only see much of a benefit here for queries that have a filter.