@stevendanna 2024-05-20
I pushed some commits that looks at the queries in the presence of stats and now the diff matches my expectations a  lot better. Namely, queries that still have to scan most of the job_info rows aren't much better, but queries with high selectivity in there where clause get a good deal better:

```
Jobs/crdb_internal.system_jobs:

|--------------------------------------------+------------------------------------------+-----------------------|
| f69856e6ae0080fd7dc94cc174d011cfa60c5188^1 | f69856e6ae0080fd7dc94cc174d011cfa60c5188 |                       |
|--------------------------------------------+------------------------------------------+-----------------------|
| sec/op                                     | sec/op                                   | vs base               |
| 72.36m ± 14%                               | 73.75m ± 7%                              | ~ (p=0.796 n=10)      |
|--------------------------------------------+------------------------------------------+-----------------------|
| roundtrips                                 | roundtrips                               | vs base               |
| 3.000 ± 0%                                 | 3.000 ± 0%                               | ~ (p=1.000 n=10)      |
|--------------------------------------------+------------------------------------------+-----------------------|
| B/op                                       | B/op                                     | vs base               |
| 26.52Mi ± 2%                               | 26.03Mi ± 9%                             | -1.87% (p=0.015 n=10) |
|--------------------------------------------+------------------------------------------+-----------------------|
| allocs/op                                  | allocs/op                                | vs base               |
|                                            |                                          |                       |
| 316.0k ± 1%                                | 317.2k ± 1%                              | ~ (p=0.912 n=10)      |


                               
Jobs/jobs_page_type_filtered_no_matches

|--------------------------------------------+------------------------------------------+------------------------|
| f69856e6ae0080fd7dc94cc174d011cfa60c5188^1 | f69856e6ae0080fd7dc94cc174d011cfa60c5188 |                        |
|--------------------------------------------+------------------------------------------+------------------------|
| sec/op                                     | sec/op                                   | vs base                |
|                                            |                                          |                        |
| 70.06m ± 7%                                | 14.65m ± 7%                              | -79.09% (p=0.000 n=10) |
|--------------------------------------------+------------------------------------------+------------------------|
| roundtrips                                 | roundtrips                               | vs base                |
| 3.000 ± 0%                                 | 1.000 ± 0%                               | -66.67% (p=0.000 n=10) |
|--------------------------------------------+------------------------------------------+------------------------|
| B/op                                       | B/op                                     | vs base                |
| 8.215Mi ± 8%                               | 3.633Mi ± 19%                            | -55.77% (p=0.000 n=10) |
|--------------------------------------------+------------------------------------------+------------------------|
| allocs/op                                  | allocs/op                                | vs base                |
| 90.05k ± 14%                               | 49.63k ± 22%                             | -44.88% (p=0.000 n=10) |
|--------------------------------------------+------------------------------------------+------------------------|


```