@stevendanna 2024-05-29
@dt I've removed the deeper backports tags from this given our uncertainty about how well this query fares in the absence of query statistics in older releases. 

Note that I _think_ this probably accounts for most of the variance I was seeing when trying to convert this to a virtual view. So, once this is in I think it makes sense to rebase that PR on top of this and see if the virtual view no longer produces a regression.